Add filter functionality to cabins table
Updated the UI to add filtering options to the cabins table. This allows users to filter cabins based on their discount status. Also, validation has been added to ensure that the discount value is less than or equal to the regular price of the cabin. A minor update was made to handle empty data arrays.