Implement protected routes and user authentication
A ProtectedRoute component was introduced and used to protect the app's routes. The user's session is now being retrieved from the Supabase library and stored using the "@tanstack/react-query" library in the useLogin hook. A new useUser hook was created to retrieve the user's status and determine whether the user is authenticated. Unsuccessful login attempts now result in route redirection.